Add install and uninstall commands
Users can now install plugins from github repositories with:

    $ gh install <user>/<repo>

Plugins are loaded based on a file called manifest which is a just a
file listing the plugins:

    # example manifest file
    user1/plugin1
    user1/plugin2
    user2/plugin1

The manifest is located in:

    # Ruby
    Dir.home + "/.githubgem"

Right now the only way to disable/enable plugins without
uninstalling/installing them is to manually edit the manifest.
Hopefully commands will take care of this.